@@ -24,8 +24,8 @@
     "events": "^1.1.1",
     "fetch-ie8": "^1.4.3",
     "flux": "^2.1.1",
-    "react": "^0.14.8",
-    "react-dom": "^0.14.8",
+    "react": "^15.3.0",
+    "react-dom": "^15.3.0",
     "react-router": "^2.6.1"
   },
   "devDependencies": {
