@@ -7,7 +7,7 @@ export class Filters extends React.Component {
   constructor() {
     super()
     this.state = {
-      collapsed: true,  
+      collapsed: true,
     };
   }
 
