@@ -1,11 +1,12 @@
 const React = require('react');
+export const Link = require('react-router').Link;
 
 const Filters = require('../components/layout/Filters').Filters;
 const Calendar = require('../components/layout/Calendar').Calendar;
 const CalendarActions = require('../actions/CalendarActions');
 const Header = require('../components/layout/Header').Header;
 
-export class Weekly extends React.Component {
+export default class Weekly extends React.Component {
 /*  constructor() {
     super();
     this.getTodos = this.getTodos.bind(this);
